Мелкие изменения

+++ src/main/java/ru/telegramBot/gm/app/urfuData/Group.java
@@ -1,3 +1,3 @@
 package ru.telegramBot.gm.app.urfuData;
 
-public record Group(int id, int instituteId) {}
+public record Group(long id, long instituteId) {}
